Create Sequential Builder
Summary:
This diff creates a class GraphSequentialBuilder that takes a set of targets and
builds them sequentially. Right now, when you launch falcon, it builds all the
root targets.

It is quite simple for now, there are many things that we don't handle yet with
subprocesses such as piping stdout and stderr, etc.

Test Plan: Generate a json file and run falcon on it.

Reviewers: NicolasDP

Reviewed By: NicolasDP

Differential Revision: http://91.121.199.76/D13