Create thrift service and client.
Summary:
This diff creates a basic thrift service and builds it for cpp and python.
DaemonInstance now launches a thrift server and a python client script can
connect to it.

I implement some cmake helper functions for generating thrift modules in cpp and
python as well as a cmake function for moving the required generated python
files to clients/python/

Test Plan:
Launch the falcon server then run ./clients/python/falcon.py to test.
The client script only send a request for the pid of the server.

Reviewers: NicolasDP

Reviewed By: NicolasDP

Subscribers: adrienconrath

Differential Revision: http://91.121.199.76/D15