Logging: introduce logging basis.
Actually we are using the trivial logging.
The API won't change in futre, so use:
LOG(__level__) << "message"; //without new line

The new line is totaly useless here since it will be managed by the backend.
So...

Summary:
..., if you have debug message, instead of using "std::cout <<", use "LOG(debug)
<<"

In a near future version: I will add the option to store the log in different
backends (in a file, in the error output, through the network...).

In a future version (don't know when yet) I will also use the same lib
(boost::log).

Introduce logging system basis

Test Plan: compile, use it, and see the standard output

Reviewers: adrienconrath

Reviewed By: adrienconrath

Subscribers: NicolasDP

Differential Revision: http://91.121.199.76/D18