Create sub process manager class
Summary:
This diff creates a class PosixSubprocessManager that aims at monitoring several
subprocesses. When a subprocess is spawned, stdout and stderr are piped to some
internal buffers for later use. PosixSubprocessManager uses poll to read from
those file descriptors.

This is the first needed step for the capacity to have commands run in parallel
and also be able to send the standard and error outputs of a command to a
client connected via a socket.

Test Plan:
Create a makefile with commands that trigger some output,
Run the daemon, build using the client.

Reviewers: NicolasDP

Reviewed By: NicolasDP

Subscribers: adrienconrath

Differential Revision: http://91.121.199.76/D20