Update the GraphVisitors
Summary:
Add a PrinterVisitor interface that all Visitor which provide an output HAVE TO
implement. Visit class define a std::ostream where every stream should be
directed to.

Also add color and some comments in the GraphGraphizPrinter
Fixe the GraphMakefilePrinter: do not print duplicated rules

Test Plan:
to check the fix in the makefileprinter: generate it and use it
to check the color generate the dot file "-M dot" and use "dot -Tpng
file.dot>file.png"

Reviewers: adrienconrath

Reviewed By: adrienconrath

Subscribers: NicolasDP

Differential Revision: http://91.121.199.76/D22