Create a streaming server for streaming the build output to multiple clients.
Summary:
This creates a server that accepts connections from multiple clients. When a
client connects, it receives the entire build output.

For now, it does not handle any shutdown mechanism, and the client has no way of
knowing that the build output is complete. This will be implemented in the next
diff.

Test Plan:
Run the daemon, then open several terminals to connect to the stream
server: "nc localhost 4343". Then trigger a build for which there is an output
(a compiling error for example). Check that each nc instance receives the
output.

Note: this uses poll and eventfd. The whole thing works without any threading.

Reviewers: NicolasDP

Reviewed By: NicolasDP

Subscribers: adrienconrath

Differential Revision: http://91.121.199.76/D24