Implement shutdown mechanism
Summary:
It is now possible to call the shutdown thrift command in order to shutdown the
daemon. It is also possible to only interrupt the current build.

This diff contains the shutdown logic in the daemon as well as two new options
in the python client: -s for shutting down, -i for interrupting the current
build.

Test Plan:
1) Start the daemon. 2) Start a build: falcon.py -b. 3) Shutdown the
daemon (this will interrupt the build): falcon.py -s

Reviewers: NicolasDP

Reviewed By: NicolasDP

Subscribers: adrienconrath

Maniphest Tasks: T10

Differential Revision: http://91.121.199.76/D31