Detect missing sources
Summary:
I added an expression in the watchman trigger which makes it possible to be
notified when a file is deleted as well as modified.

In addition to that, if a source file is deleted, the next time you perform a
build, the falcon client prints an error.

Test Plan:
falcon --build && rm ../src/main.cpp && falcon --build
This should display:
Error, cannot build without: ../src/main.cpp

Reviewers: NicolasDP

Reviewed By: NicolasDP

Differential Revision: http://91.121.199.76/D49